New step for grid search, average performance, val at the beginning
Added averaging of the performance, so it generates the sample 20 times and then takes the average. Then added validation at the beginning of training so the plot looks nicer. Lastly, modified parameters to perform next step of the grid search, so now we have the best performing architecture for both 1 and 2 layer lstm and rnn.